Able to send in CSV data via CSVReader

+++ Väderdata.Web/Data/CsvModelClass.cs
@@ -11,6 +11,6 @@ namespace Väderdata.Web.Data
         public DateTime Datum { get; set; }
         public string Plats { get; set; }
         public double Temp { get; set; }
-        public double Luftfuktighet { get; set; }
+        public int Luftfuktighet { get; set; }
     }
 }
